requirements.txt
streamlit
google-generativeai
agno-agents
duckduckgo-search
